ProjectsPage: simplify code
