+++ src/Pages/Projects/Components/Link-Github-Project/Link-Github-Project.js
@@ -1,4 +1,5 @@
 import React, {Component} from 'react';
+import {sourceCodeName} from './assets/language';
 
 // css
 import './Link-Github-Project.css';
@@ -7,14 +8,6 @@ import './Link-Github-Project.css';
 import GithubLogo from './logo/github.svg';
 
 class LinkGithubProject extends Component {
-  sourceCodeName() {
-    if (localStorage.getItem('language') === 'FR') {
-      return ('Code Source');
-    } else {
-      return ('Source Code');
-    }
-  }
-
   render() {
     return (
       <div className="repo-github-link">
@@ -22,7 +15,7 @@ class LinkGithubProject extends Component {
           <button type="button" className="repo-github-link-button">
             <img className="repo-github-link-button_ico" src={GithubLogo} alt="github logo" />
             <span className="repo-github-link-button_text">
-              {this.sourceCodeName()}
+              {sourceCodeName()}
             </span>
           </button>
         </a>
