+++ src/Pages/Projects/Components/Link-Github-Project/assets/language.js
@@ -1,0 +1,9 @@
+import {flagsSelecting} from '../../../../../Language/flags-selector';
+
+export const sourceCodeName = () => {
+  if (flagsSelecting() === 'fr') {
+    return ('Code Source');
+  } else {
+    return ('Source Code');
+  }
+};
